Former-commit-id: 38ff2d270e0c1b3f92a7bb7a10dc0a87e02121f0

+++ BACK/muscle_maker/src/main/java/com/ssafy/muscle_maker/dto/inBody/request/CreateInBodyRequest.java
@@ -7,8 +7,8 @@ import lombok.*;
 @Builder
 public class CreateInBodyRequest {
 
-    private Float weight;
-    private Float muscleMass;
-    private Float fatMass;
+    private Double weight;
+    private Double muscleMass;
+    private Double fatMass;
 
 }
